Consolidate RAG modules into unified implementation
