--- core/rag/cognitive_nexus.py
@@ -1,1085 +1,0 @@
-"""
-Cognitive Nexus - Advanced Analysis and Reasoning Engine
-
-Multi-perspective reasoning system for retrieved information:
-- Cross-reference validation between sources
-- Uncertainty quantification and confidence scoring
-- Contextual relevance assessment
-- Synthesis of multiple information sources
-- Detection of contradictions and inconsistencies
-- Belief propagation with Bayesian networks
-- Meta-cognitive reasoning about reasoning processes
-
-Integrated with fog computing for distributed analysis.
-"""
-
-import asyncio
-from dataclasses import dataclass, field
-from enum import Enum
-import logging
-import time
-from typing import Any
-
-import numpy as np
-
-logger = logging.getLogger(__name__)
-
-
-class AnalysisType(Enum):
-    """Types of cognitive analysis that can be performed."""
-
-    FACTUAL_VERIFICATION = "factual_verification"
-    CONSISTENCY_CHECK = "consistency_check"
-    RELEVANCE_ASSESSMENT = "relevance_assessment"
-    UNCERTAINTY_QUANTIFICATION = "uncertainty_quantification"
-    SYNTHESIS = "synthesis"
-    CONTRADICTION_DETECTION = "contradiction_detection"
-    BELIEF_PROPAGATION = "belief_propagation"
-    META_REASONING = "meta_reasoning"
-
-
-class ConfidenceLevel(Enum):
-    """Confidence levels for analysis results."""
-
-    VERY_LOW = 0.1
-    LOW = 0.3
-    MODERATE = 0.5
-    HIGH = 0.7
-    VERY_HIGH = 0.9
-
-
-class ReasoningStrategy(Enum):
-    """Different reasoning strategies for analysis."""
-
-    DEDUCTIVE = "deductive"  # From general principles to specific conclusions
-    INDUCTIVE = "inductive"  # From specific observations to general patterns
-    ABDUCTIVE = "abductive"  # Best explanation for observations
-    ANALOGICAL = "analogical"  # Reasoning by analogy and similarity
-    PROBABILISTIC = "probabilistic"  # Bayesian probabilistic reasoning
-
-
-@dataclass
-class RetrievedInformation:
-    """Information retrieved from RAG system with enhanced metadata."""
-
-    id: str
-    content: str
-    source: str
-    relevance_score: float
-    retrieval_confidence: float
-
-    # Context tags from dual context system
-    book_summary: str = ""
-    chapter_summary: str = ""
-
-    # Vector store metadata
-    embedding_similarity: float = 0.0
-    chunk_index: int = 0
-
-    # Graph RAG metadata
-    graph_connections: list[str] = field(default_factory=list)
-    relationship_types: list[str] = field(default_factory=list)
-    trust_score: float = 0.5
-    centrality_score: float = 0.0
-
-    # Episodic metadata
-    recency_score: float = 0.0
-    access_frequency: int = 0
-    decay_applied: bool = False
-
-    timestamp: float = field(default_factory=time.time)
-
-
-@dataclass
-class BeliefNode:
-    """Node in Bayesian belief network for probabilistic reasoning."""
-
-    id: str
-    statement: str
-    prior_probability: float
-    current_probability: float
-
-    # Evidence supporting/contradicting this belief
-    supporting_evidence: list[str] = field(default_factory=list)
-    contradicting_evidence: list[str] = field(default_factory=list)
-
-    # Dependencies and influences
-    parent_beliefs: list[str] = field(default_factory=list)
-    child_beliefs: list[str] = field(default_factory=list)
-
-    # Metadata
-    confidence: float = 0.5
-    last_updated: float = field(default_factory=time.time)
-    update_count: int = 0
-
-
-@dataclass
-class AnalysisResult:
-    """Result of cognitive analysis with detailed reasoning trace."""
-
-    analysis_type: AnalysisType
-    confidence: ConfidenceLevel
-    result: dict[str, Any]
-    reasoning: str
-    strategy_used: ReasoningStrategy
-
-    # Supporting information
-    sources_analyzed: list[str] = field(default_factory=list)
-    contradictions_found: list[dict[str, Any]] = field(default_factory=list)
-    uncertainties: list[str] = field(default_factory=list)
-
-    # Belief network updates
-    belief_updates: list[dict[str, Any]] = field(default_factory=list)
-
-    # Meta-reasoning
-    reasoning_trace: list[str] = field(default_factory=list)
-    alternative_interpretations: list[str] = field(default_factory=list)
-
-    analysis_duration_ms: float = 0.0
-    timestamp: float = field(default_factory=time.time)
-
-
-@dataclass
-class SynthesizedAnswer:
-    """Final synthesized answer with comprehensive analysis."""
-
-    answer: str
-    confidence: float
-    supporting_sources: list[str]
-
-    # Analysis components
-    factual_accuracy: float = 0.0
-    consistency_score: float = 0.0
-    relevance_score: float = 0.0
-    completeness_score: float = 0.0
-
-    # Uncertainty indicators
-    conflicting_information: list[str] = field(default_factory=list)
-    knowledge_gaps: list[str] = field(default_factory=list)
-    reliability_concerns: list[str] = field(default_factory=list)
-
-    # Reasoning details
-    synthesis_method: str = ""
-    reasoning_strategy: ReasoningStrategy = ReasoningStrategy.PROBABILISTIC
-    alternative_answers: list[str] = field(default_factory=list)
-
-    # Belief network state
-    key_beliefs: list[BeliefNode] = field(default_factory=list)
-    belief_confidence: float = 0.5
-
-    creation_timestamp: float = field(default_factory=time.time)
-
-
-class CognitiveNexus:
-    """
-    Advanced Cognitive Analysis System
-
-    Performs sophisticated multi-perspective analysis of retrieved information:
-    - Bayesian belief propagation for probabilistic reasoning
-    - Meta-cognitive analysis of reasoning processes
-    - Cross-source consistency and contradiction detection
-    - Uncertainty quantification with confidence bounds
-    - Multi-strategy reasoning (deductive, inductive, abductive, analogical)
-    - Integration with fog computing for distributed analysis
-    """
-
-    def __init__(self, enable_fog_computing: bool = False):
-        """Initialize Cognitive Nexus with optional fog computing."""
-
-        # Analysis configuration
-        self.confidence_threshold = 0.7
-        self.max_sources_per_synthesis = 15
-        self.contradiction_sensitivity = 0.3
-        self.relevance_threshold = 0.5
-        self.belief_propagation_iterations = 10
-        self.meta_reasoning_depth = 3
-
-        # Belief network for probabilistic reasoning
-        self.belief_network: dict[str, BeliefNode] = {}
-        self.belief_dependencies: dict[str, list[str]] = {}
-
-        # Analysis history and caching
-        self.analysis_history: list[AnalysisResult] = []
-        self.synthesis_cache: dict[str, SynthesizedAnswer] = {}
-        self.reasoning_patterns: dict[str, dict[str, Any]] = {}
-
-        # Performance tracking
-        self.stats = {
-            "analyses_performed": 0,
-            "synthesis_requests": 0,
-            "contradictions_detected": 0,
-            "uncertainties_identified": 0,
-            "belief_updates": 0,
-            "fog_tasks_executed": 0,
-            "meta_reasoning_cycles": 0,
-        }
-
-        # Fog computing integration
-        self.enable_fog_computing = enable_fog_computing
-        self.fog_workers: list[Any] = []
-
-        self.initialized = False
-
-    async def initialize(self):
-        """Initialize the Cognitive Nexus analysis system."""
-        try:
-            logger.info("Initializing Cognitive Nexus analysis system...")
-
-            # Initialize belief network with common reasoning patterns
-            await self._initialize_belief_network()
-
-            # Start background analysis optimization
-            asyncio.create_task(self._periodic_optimization())
-
-            # Initialize fog computing workers if enabled
-            if self.enable_fog_computing:
-                await self._initialize_fog_workers()
-
-            self.initialized = True
-            logger.info("✅ Cognitive Nexus initialized successfully")
-
-        except Exception as e:
-            logger.error(f"❌ Cognitive Nexus initialization failed: {e}")
-            raise
-
-    async def analyze_retrieved_information(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        analysis_types: list[AnalysisType] | None = None,
-        reasoning_strategy: ReasoningStrategy = ReasoningStrategy.PROBABILISTIC,
-    ) -> list[AnalysisResult]:
-        """
-        Perform comprehensive cognitive analysis of retrieved information.
-
-        Uses multiple reasoning strategies and integrates with belief networks
-        for sophisticated multi-perspective analysis.
-        """
-        if not self.initialized:
-            raise RuntimeError("CognitiveNexus not initialized")
-
-        if not analysis_types:
-            analysis_types = [
-                AnalysisType.FACTUAL_VERIFICATION,
-                AnalysisType.CONSISTENCY_CHECK,
-                AnalysisType.RELEVANCE_ASSESSMENT,
-                AnalysisType.UNCERTAINTY_QUANTIFICATION,
-                AnalysisType.BELIEF_PROPAGATION,
-            ]
-
-        start_time = time.time()
-        results = []
-
-        logger.info(f"Analyzing {len(retrieved_info)} sources for query: '{query[:50]}...'")
-        logger.info(f"Using reasoning strategy: {reasoning_strategy.value}")
-
-        # Parallel analysis if fog computing is enabled
-        if self.enable_fog_computing and len(retrieved_info) > 10:
-            results = await self._distributed_analysis(query, retrieved_info, analysis_types, reasoning_strategy)
-        else:
-            # Sequential analysis for smaller datasets
-            for analysis_type in analysis_types:
-                try:
-                    result = await self._perform_analysis(query, retrieved_info, analysis_type, reasoning_strategy)
-                    result.analysis_duration_ms = (time.time() - start_time) * 1000
-                    results.append(result)
-
-                    self.stats["analyses_performed"] += 1
-
-                except Exception as e:
-                    logger.error(f"Analysis failed for {analysis_type.value}: {e}")
-
-        # Perform meta-reasoning on analysis results
-        meta_result = await self._meta_reasoning_analysis(query, retrieved_info, results)
-        if meta_result:
-            results.append(meta_result)
-            self.stats["meta_reasoning_cycles"] += 1
-
-        # Update belief network based on new evidence
-        await self._update_belief_network(query, retrieved_info, results)
-
-        # Store analysis history
-        self.analysis_history.extend(results)
-        if len(self.analysis_history) > 1000:
-            self.analysis_history = self.analysis_history[-1000:]
-
-        total_time = (time.time() - start_time) * 1000
-        logger.info(f"Cognitive analysis completed in {total_time:.1f}ms")
-
-        return results
-
-    async def synthesize_answer(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        analysis_results: list[AnalysisResult] | None = None,
-        reasoning_strategy: ReasoningStrategy = ReasoningStrategy.PROBABILISTIC,
-    ) -> SynthesizedAnswer:
-        """
-        Synthesize comprehensive answer using cognitive analysis and belief networks.
-
-        Integrates multiple reasoning strategies and belief propagation for
-        sophisticated answer synthesis with uncertainty quantification.
-        """
-        synthesis_key = f"query_{hash(query)}_{len(retrieved_info)}_{reasoning_strategy.value}"
-
-        # Check cache
-        if synthesis_key in self.synthesis_cache:
-            cached = self.synthesis_cache[synthesis_key]
-            if time.time() - cached.creation_timestamp < 3600:  # 1 hour cache
-                return cached
-
-        start_time = time.time()
-
-        # Perform analysis if not provided
-        if not analysis_results:
-            analysis_results = await self.analyze_retrieved_information(
-                query, retrieved_info, reasoning_strategy=reasoning_strategy
-            )
-
-        # Extract analysis insights
-        factual_score = self._extract_analysis_score(analysis_results, AnalysisType.FACTUAL_VERIFICATION)
-        consistency_score = self._extract_analysis_score(analysis_results, AnalysisType.CONSISTENCY_CHECK)
-        relevance_score = self._extract_analysis_score(analysis_results, AnalysisType.RELEVANCE_ASSESSMENT)
-
-        # Find contradictions and uncertainties
-        contradictions = self._find_contradictions(retrieved_info, analysis_results)
-        uncertainties = self._identify_uncertainties(analysis_results)
-        knowledge_gaps = self._identify_knowledge_gaps(query, retrieved_info)
-
-        # Apply reasoning strategy to synthesis
-        synthesized_text = await self._generate_synthesized_text(
-            query, retrieved_info, analysis_results, reasoning_strategy
-        )
-
-        # Generate alternative interpretations
-        alternatives = await self._generate_alternative_answers(
-            query, retrieved_info, analysis_results, reasoning_strategy
-        )
-
-        # Extract relevant beliefs from network
-        key_beliefs = self._extract_relevant_beliefs(query, retrieved_info)
-        belief_confidence = self._calculate_belief_confidence(key_beliefs)
-
-        # Calculate overall confidence using multiple factors
-        confidence = self._calculate_synthesis_confidence(
-            factual_score,
-            consistency_score,
-            relevance_score,
-            len(contradictions),
-            belief_confidence,
-            reasoning_strategy,
-        )
-
-        # Assess completeness
-        completeness = await self._assess_answer_completeness(query, synthesized_text, retrieved_info, key_beliefs)
-
-        # Create synthesized answer
-        answer = SynthesizedAnswer(
-            answer=synthesized_text,
-            confidence=confidence,
-            supporting_sources=[info.source for info in retrieved_info],
-            factual_accuracy=factual_score,
-            consistency_score=consistency_score,
-            relevance_score=relevance_score,
-            completeness_score=completeness,
-            conflicting_information=contradictions,
-            knowledge_gaps=knowledge_gaps,
-            reliability_concerns=uncertainties,
-            synthesis_method=f"cognitive_nexus_v3_{reasoning_strategy.value}",
-            reasoning_strategy=reasoning_strategy,
-            alternative_answers=alternatives,
-            key_beliefs=key_beliefs,
-            belief_confidence=belief_confidence,
-            creation_timestamp=time.time(),
-        )
-
-        # Cache result
-        self.synthesis_cache[synthesis_key] = answer
-        self.stats["synthesis_requests"] += 1
-
-        duration = (time.time() - start_time) * 1000
-        logger.info(f"Synthesis completed in {duration:.1f}ms (confidence: {confidence:.3f})")
-
-        return answer
-
-    async def _perform_analysis(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        analysis_type: AnalysisType,
-        reasoning_strategy: ReasoningStrategy,
-    ) -> AnalysisResult:
-        """Perform specific type of cognitive analysis."""
-
-        reasoning_trace = [f"Starting {analysis_type.value} analysis using {reasoning_strategy.value} reasoning"]
-
-        if analysis_type == AnalysisType.FACTUAL_VERIFICATION:
-            return await self._verify_factual_accuracy(query, retrieved_info, reasoning_strategy, reasoning_trace)
-        elif analysis_type == AnalysisType.CONSISTENCY_CHECK:
-            return await self._check_consistency(retrieved_info, reasoning_strategy, reasoning_trace)
-        elif analysis_type == AnalysisType.RELEVANCE_ASSESSMENT:
-            return await self._assess_relevance(query, retrieved_info, reasoning_strategy, reasoning_trace)
-        elif analysis_type == AnalysisType.UNCERTAINTY_QUANTIFICATION:
-            return await self._quantify_uncertainties(retrieved_info, reasoning_strategy, reasoning_trace)
-        elif analysis_type == AnalysisType.CONTRADICTION_DETECTION:
-            return await self._detect_contradictions(retrieved_info, reasoning_strategy, reasoning_trace)
-        elif analysis_type == AnalysisType.BELIEF_PROPAGATION:
-            return await self._belief_propagation_analysis(query, retrieved_info, reasoning_strategy, reasoning_trace)
-        elif analysis_type == AnalysisType.META_REASONING:
-            return await self._meta_reasoning_analysis(query, retrieved_info, [], reasoning_trace)
-        else:
-            raise ValueError(f"Unknown analysis type: {analysis_type}")
-
-    async def _verify_factual_accuracy(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        reasoning_strategy: ReasoningStrategy,
-        reasoning_trace: list[str],
-    ) -> AnalysisResult:
-        """Verify factual accuracy using specified reasoning strategy."""
-
-        fact_scores = []
-        reasoning_trace.append("Evaluating factual accuracy of sources")
-
-        for info in retrieved_info:
-            # Assess source reliability
-            source_reliability = self._assess_source_reliability(info)
-
-            # Count factual indicators based on reasoning strategy
-            if reasoning_strategy == ReasoningStrategy.DEDUCTIVE:
-                factual_indicators = self._count_deductive_indicators(info.content)
-            elif reasoning_strategy == ReasoningStrategy.INDUCTIVE:
-                factual_indicators = self._count_inductive_indicators(info.content)
-            elif reasoning_strategy == ReasoningStrategy.PROBABILISTIC:
-                factual_indicators = self._count_probabilistic_indicators(info.content)
-            else:
-                factual_indicators = self._count_factual_indicators(info.content)
-
-            # Combine scores with strategy weighting
-            fact_score = self._combine_scores_by_strategy(source_reliability, factual_indicators, reasoning_strategy)
-            fact_scores.append(fact_score)
-
-            reasoning_trace.append(
-                f"Source {info.id}: reliability={source_reliability:.3f}, indicators={factual_indicators:.3f}"
-            )
-
-        avg_accuracy = np.mean(fact_scores) if fact_scores else 0.0
-        confidence = ConfidenceLevel.HIGH if avg_accuracy > 0.8 else ConfidenceLevel.MODERATE
-
-        reasoning_trace.append(f"Average factual accuracy: {avg_accuracy:.3f}")
-
-        return AnalysisResult(
-            analysis_type=AnalysisType.FACTUAL_VERIFICATION,
-            confidence=confidence,
-            result={
-                "overall_accuracy": avg_accuracy,
-                "source_scores": fact_scores,
-                "reliable_sources": sum(1 for score in fact_scores if score > 0.7),
-                "strategy_used": reasoning_strategy.value,
-            },
-            reasoning=f"Analyzed {len(retrieved_info)} sources with {reasoning_strategy.value} reasoning, avg accuracy {avg_accuracy:.3f}",
-            strategy_used=reasoning_strategy,
-            sources_analyzed=[info.id for info in retrieved_info],
-            reasoning_trace=reasoning_trace.copy(),
-        )
-
-    async def _belief_propagation_analysis(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        reasoning_strategy: ReasoningStrategy,
-        reasoning_trace: list[str],
-    ) -> AnalysisResult:
-        """Perform belief propagation analysis on retrieved information."""
-
-        reasoning_trace.append("Starting belief propagation analysis")
-
-        # Extract beliefs from retrieved information
-        extracted_beliefs = []
-        for info in retrieved_info:
-            beliefs = self._extract_beliefs_from_content(info.content, info.id)
-            extracted_beliefs.extend(beliefs)
-            reasoning_trace.append(f"Extracted {len(beliefs)} beliefs from {info.id}")
-
-        # Update belief network with new evidence
-        belief_updates = []
-        for belief in extracted_beliefs:
-            if belief["id"] in self.belief_network:
-                # Update existing belief
-                old_prob = self.belief_network[belief["id"]].current_probability
-                new_prob = self._update_belief_probability(belief, retrieved_info)
-                self.belief_network[belief["id"]].current_probability = new_prob
-                self.belief_network[belief["id"]].update_count += 1
-
-                belief_updates.append(
-                    {
-                        "belief_id": belief["id"],
-                        "old_probability": old_prob,
-                        "new_probability": new_prob,
-                        "evidence_strength": belief.get("confidence", 0.5),
-                    }
-                )
-
-                reasoning_trace.append(f"Updated belief {belief['id']}: {old_prob:.3f} -> {new_prob:.3f}")
-            else:
-                # Add new belief
-                new_belief = BeliefNode(
-                    id=belief["id"],
-                    statement=belief["statement"],
-                    prior_probability=belief.get("prior", 0.5),
-                    current_probability=belief.get("confidence", 0.5),
-                    supporting_evidence=[
-                        info.id for info in retrieved_info if belief["statement"].lower() in info.content.lower()
-                    ],
-                )
-                self.belief_network[belief["id"]] = new_belief
-
-                belief_updates.append(
-                    {"belief_id": belief["id"], "action": "created", "probability": new_belief.current_probability}
-                )
-
-                reasoning_trace.append(
-                    f"Created new belief {belief['id']} with probability {new_belief.current_probability:.3f}"
-                )
-
-        # Propagate beliefs through network
-        propagation_iterations = min(self.belief_propagation_iterations, len(self.belief_network))
-        for i in range(propagation_iterations):
-            changes = self._propagate_beliefs_one_iteration()
-            if changes < 0.001:  # Convergence threshold
-                reasoning_trace.append(f"Belief propagation converged after {i+1} iterations")
-                break
-
-        self.stats["belief_updates"] += len(belief_updates)
-
-        # Calculate overall belief confidence
-        relevant_beliefs = [
-            b
-            for b in self.belief_network.values()
-            if any(term in b.statement.lower() for term in query.lower().split())
-        ]
-
-        belief_confidence = np.mean([b.current_probability for b in relevant_beliefs]) if relevant_beliefs else 0.5
-
-        return AnalysisResult(
-            analysis_type=AnalysisType.BELIEF_PROPAGATION,
-            confidence=ConfidenceLevel.HIGH if belief_confidence > 0.7 else ConfidenceLevel.MODERATE,
-            result={
-                "belief_updates": len(belief_updates),
-                "belief_confidence": belief_confidence,
-                "relevant_beliefs": len(relevant_beliefs),
-                "propagation_iterations": propagation_iterations,
-                "network_size": len(self.belief_network),
-            },
-            reasoning=f"Updated {len(belief_updates)} beliefs, network confidence: {belief_confidence:.3f}",
-            strategy_used=reasoning_strategy,
-            sources_analyzed=[info.id for info in retrieved_info],
-            belief_updates=belief_updates,
-            reasoning_trace=reasoning_trace.copy(),
-        )
-
-    # Helper methods (implementing core cognitive functions)
-
-    def _assess_source_reliability(self, info: RetrievedInformation) -> float:
-        """Assess reliability of information source with enhanced metrics."""
-        base_score = info.retrieval_confidence
-
-        # Trust score from graph RAG
-        if hasattr(info, "trust_score") and info.trust_score > 0:
-            base_score = (base_score + info.trust_score) / 2
-
-        # Boost for academic/authoritative sources
-        authoritative_terms = ["academic", "journal", "university", "gov", "peer-reviewed", "study", "research"]
-        if any(term in info.source.lower() for term in authoritative_terms):
-            base_score += 0.2
-
-        # Boost for graph centrality
-        if hasattr(info, "centrality_score") and info.centrality_score > 0:
-            base_score += info.centrality_score * 0.1
-
-        # Penalty for low recency (if available)
-        if hasattr(info, "recency_score") and info.recency_score < 0.3:
-            base_score -= 0.1
-
-        return min(1.0, max(0.0, base_score))
-
-    def _count_factual_indicators(self, content: str) -> float:
-        """Count indicators of factual content."""
-        factual_indicators = [
-            "study",
-            "research",
-            "data",
-            "statistics",
-            "evidence",
-            "proven",
-            "demonstrated",
-            "measured",
-            "observed",
-            "experiment",
-            "analysis",
-            "findings",
-            "results",
-            "conclude",
-            "significant",
-            "correlation",
-        ]
-
-        count = sum(1 for indicator in factual_indicators if indicator in content.lower())
-        return min(1.0, count / 8)  # Normalize to 0-1
-
-    def _count_deductive_indicators(self, content: str) -> float:
-        """Count indicators of deductive reasoning patterns."""
-        deductive_indicators = [
-            "therefore",
-            "thus",
-            "hence",
-            "consequently",
-            "follows that",
-            "given that",
-            "since",
-            "because",
-            "if...then",
-            "logical",
-        ]
-
-        count = sum(1 for indicator in deductive_indicators if indicator in content.lower())
-        return min(1.0, count / 5)
-
-    def _count_inductive_indicators(self, content: str) -> float:
-        """Count indicators of inductive reasoning patterns."""
-        inductive_indicators = [
-            "pattern",
-            "trend",
-            "generally",
-            "typically",
-            "often",
-            "usually",
-            "suggests",
-            "indicates",
-            "appears",
-            "tends to",
-            "commonly",
-        ]
-
-        count = sum(1 for indicator in inductive_indicators if indicator in content.lower())
-        return min(1.0, count / 5)
-
-    def _count_probabilistic_indicators(self, content: str) -> float:
-        """Count indicators of probabilistic reasoning."""
-        probabilistic_indicators = [
-            "likely",
-            "probably",
-            "possibly",
-            "chance",
-            "probability",
-            "risk",
-            "uncertain",
-            "confidence",
-            "estimate",
-            "approximately",
-            "around",
-        ]
-
-        count = sum(1 for indicator in probabilistic_indicators if indicator in content.lower())
-        return min(1.0, count / 5)
-
-    def _combine_scores_by_strategy(self, reliability: float, indicators: float, strategy: ReasoningStrategy) -> float:
-        """Combine scores based on reasoning strategy."""
-
-        if strategy == ReasoningStrategy.DEDUCTIVE:
-            # Emphasize logical consistency and reliability
-            return reliability * 0.7 + indicators * 0.3
-        elif strategy == ReasoningStrategy.INDUCTIVE:
-            # Emphasize patterns and evidence
-            return reliability * 0.4 + indicators * 0.6
-        elif strategy == ReasoningStrategy.PROBABILISTIC:
-            # Balanced approach with uncertainty handling
-            return reliability * 0.5 + indicators * 0.5
-        else:
-            # Default balanced combination
-            return (reliability + indicators) / 2
-
-    async def _check_consistency(
-        self,
-        retrieved_info: list[RetrievedInformation],
-        reasoning_strategy: ReasoningStrategy,
-        reasoning_trace: list[str],
-    ) -> AnalysisResult:
-        """Check consistency between sources using specified reasoning strategy."""
-
-        inconsistencies = []
-        consistency_scores = []
-
-        reasoning_trace.append(f"Checking consistency using {reasoning_strategy.value} approach")
-
-        # Compare each source with every other source
-        for i, info1 in enumerate(retrieved_info):
-            for _j, info2 in enumerate(retrieved_info[i + 1 :], i + 1):
-                if reasoning_strategy == ReasoningStrategy.DEDUCTIVE:
-                    consistency = await self._calculate_logical_consistency(info1.content, info2.content)
-                elif reasoning_strategy == ReasoningStrategy.PROBABILISTIC:
-                    consistency = await self._calculate_probabilistic_consistency(info1.content, info2.content)
-                else:
-                    consistency = await self._calculate_content_consistency(info1.content, info2.content)
-
-                consistency_scores.append(consistency)
-
-                if consistency < self.contradiction_sensitivity:
-                    inconsistencies.append(
-                        {
-                            "source1": info1.id,
-                            "source2": info2.id,
-                            "consistency_score": consistency,
-                            "conflict_type": f"{reasoning_strategy.value}_contradiction",
-                            "details": f"Inconsistency detected using {reasoning_strategy.value} reasoning",
-                        }
-                    )
-
-                reasoning_trace.append(f"Consistency {info1.id} vs {info2.id}: {consistency:.3f}")
-
-        avg_consistency = np.mean(consistency_scores) if consistency_scores else 1.0
-        confidence = ConfidenceLevel.HIGH if avg_consistency > 0.7 else ConfidenceLevel.MODERATE
-
-        if inconsistencies:
-            self.stats["contradictions_detected"] += len(inconsistencies)
-            reasoning_trace.append(f"Found {len(inconsistencies)} inconsistencies")
-
-        return AnalysisResult(
-            analysis_type=AnalysisType.CONSISTENCY_CHECK,
-            confidence=confidence,
-            result={
-                "consistency_score": avg_consistency,
-                "inconsistencies_found": len(inconsistencies),
-                "pairwise_scores": consistency_scores,
-                "strategy_used": reasoning_strategy.value,
-            },
-            reasoning=f"Found {len(inconsistencies)} inconsistencies using {reasoning_strategy.value} reasoning",
-            strategy_used=reasoning_strategy,
-            contradictions_found=inconsistencies,
-            sources_analyzed=[info.id for info in retrieved_info],
-            reasoning_trace=reasoning_trace.copy(),
-        )
-
-    async def _assess_relevance(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        reasoning_strategy: ReasoningStrategy,
-        reasoning_trace: list[str],
-    ) -> AnalysisResult:
-        """Assess relevance using specified reasoning strategy."""
-
-        relevance_scores = []
-        reasoning_trace.append(f"Assessing relevance using {reasoning_strategy.value} strategy")
-
-        for info in retrieved_info:
-            # Base relevance from retrieval system
-            base_relevance = info.relevance_score
-
-            # Enhanced relevance assessment based on strategy
-            if reasoning_strategy == ReasoningStrategy.ANALOGICAL:
-                query_overlap = await self._calculate_analogical_relevance(query, info.content)
-            elif reasoning_strategy == ReasoningStrategy.PROBABILISTIC:
-                query_overlap = await self._calculate_probabilistic_relevance(query, info.content)
-            else:
-                query_overlap = await self._calculate_query_overlap(query, info.content)
-
-            context_relevance = self._assess_context_relevance(query, info)
-
-            # Strategy-specific weighting
-            if reasoning_strategy == ReasoningStrategy.DEDUCTIVE:
-                enhanced_relevance = base_relevance * 0.6 + query_overlap * 0.3 + context_relevance * 0.1
-            elif reasoning_strategy == ReasoningStrategy.INDUCTIVE:
-                enhanced_relevance = base_relevance * 0.3 + query_overlap * 0.4 + context_relevance * 0.3
-            else:
-                enhanced_relevance = (base_relevance + query_overlap + context_relevance) / 3
-
-            relevance_scores.append(enhanced_relevance)
-            reasoning_trace.append(f"Relevance for {info.id}: {enhanced_relevance:.3f}")
-
-        avg_relevance = np.mean(relevance_scores) if relevance_scores else 0.0
-        high_relevance_count = sum(1 for score in relevance_scores if score > self.relevance_threshold)
-
-        confidence = ConfidenceLevel.HIGH if avg_relevance > 0.7 else ConfidenceLevel.MODERATE
-
-        return AnalysisResult(
-            analysis_type=AnalysisType.RELEVANCE_ASSESSMENT,
-            confidence=confidence,
-            result={
-                "average_relevance": avg_relevance,
-                "relevance_scores": relevance_scores,
-                "highly_relevant_sources": high_relevance_count,
-                "relevance_threshold": self.relevance_threshold,
-                "strategy_used": reasoning_strategy.value,
-            },
-            reasoning=f"{high_relevance_count}/{len(retrieved_info)} sources highly relevant using {reasoning_strategy.value}",
-            strategy_used=reasoning_strategy,
-            sources_analyzed=[info.id for info in retrieved_info],
-            reasoning_trace=reasoning_trace.copy(),
-        )
-
-    # Additional helper methods for sophisticated analysis...
-
-    async def _initialize_belief_network(self):
-        """Initialize belief network with common reasoning patterns."""
-
-        # Common beliefs about information reliability
-        self.belief_network["academic_reliable"] = BeliefNode(
-            id="academic_reliable",
-            statement="Academic sources are generally reliable",
-            prior_probability=0.8,
-            current_probability=0.8,
-        )
-
-        self.belief_network["peer_review_quality"] = BeliefNode(
-            id="peer_review_quality",
-            statement="Peer-reviewed content has higher quality",
-            prior_probability=0.85,
-            current_probability=0.85,
-        )
-
-        self.belief_network["recent_more_relevant"] = BeliefNode(
-            id="recent_more_relevant",
-            statement="More recent information is more relevant",
-            prior_probability=0.7,
-            current_probability=0.7,
-        )
-
-        # Set up dependencies
-        self.belief_dependencies["academic_reliable"] = ["peer_review_quality"]
-        self.belief_dependencies["peer_review_quality"] = []
-        self.belief_dependencies["recent_more_relevant"] = []
-
-        logger.info("Initialized belief network with core reasoning patterns")
-
-    def _extract_beliefs_from_content(self, content: str, source_id: str) -> list[dict[str, Any]]:
-        """Extract belief statements from content."""
-
-        beliefs = []
-
-        # Simple pattern matching for belief extraction
-        # In practice, this would use more sophisticated NLP
-        belief_patterns = [
-            "is",
-            "are",
-            "will be",
-            "should be",
-            "must be",
-            "appears to be",
-            "seems to",
-            "indicates that",
-            "suggests that",
-            "proves that",
-        ]
-
-        sentences = content.split(".")
-        for i, sentence in enumerate(sentences):
-            sentence = sentence.strip()
-            if len(sentence) > 20 and any(pattern in sentence.lower() for pattern in belief_patterns):
-                belief_id = f"{source_id}_belief_{i}"
-                beliefs.append(
-                    {
-                        "id": belief_id,
-                        "statement": sentence,
-                        "confidence": 0.6,  # Default confidence
-                        "prior": 0.5,
-                        "source": source_id,
-                    }
-                )
-
-        return beliefs[:5]  # Limit to top 5 beliefs per source
-
-    def _update_belief_probability(self, belief: dict[str, Any], evidence: list[RetrievedInformation]) -> float:
-        """Update belief probability based on new evidence."""
-
-        # Simple Bayesian update (in practice would be more sophisticated)
-        prior = belief.get("prior", 0.5)
-        likelihood = belief.get("confidence", 0.5)
-
-        # Evidence strength from supporting sources
-        supporting_sources = sum(1 for info in evidence if belief["statement"].lower() in info.content.lower())
-        evidence_strength = min(1.0, supporting_sources / len(evidence))
-
-        # Bayesian update
-        posterior = (likelihood * evidence_strength * prior) / (
-            (likelihood * evidence_strength * prior) + ((1 - likelihood) * (1 - evidence_strength) * (1 - prior))
-        )
-
-        return min(0.95, max(0.05, posterior))  # Bounded between 0.05 and 0.95
-
-    def _propagate_beliefs_one_iteration(self) -> float:
-        """Perform one iteration of belief propagation."""
-
-        total_change = 0.0
-
-        for belief_id, belief in self.belief_network.items():
-            if belief_id in self.belief_dependencies:
-                # Update based on parent beliefs
-                parent_beliefs = [
-                    self.belief_network[parent_id]
-                    for parent_id in self.belief_dependencies[belief_id]
-                    if parent_id in self.belief_network
-                ]
-
-                if parent_beliefs:
-                    # Simple influence propagation
-                    old_prob = belief.current_probability
-                    parent_influence = np.mean([b.current_probability for b in parent_beliefs])
-
-                    # Weighted update
-                    new_prob = 0.8 * old_prob + 0.2 * parent_influence
-                    belief.current_probability = new_prob
-
-                    total_change += abs(new_prob - old_prob)
-
-        return total_change
-
-    async def _meta_reasoning_analysis(
-        self,
-        query: str,
-        retrieved_info: list[RetrievedInformation],
-        analysis_results: list[AnalysisResult],
-        reasoning_trace: list[str] | None = None,
-    ) -> AnalysisResult | None:
-        """Perform meta-reasoning about the reasoning process itself."""
-
-        if reasoning_trace is None:
-            reasoning_trace = ["Starting meta-reasoning analysis"]
-
-        # Analyze the quality of the reasoning process
-        reasoning_quality_indicators = []
-
-        # Check for reasoning strategy diversity
-        strategies_used = set()
-        for result in analysis_results:
-            if hasattr(result, "strategy_used"):
-                strategies_used.add(result.strategy_used.value)
-
-        strategy_diversity = len(strategies_used) / len(ReasoningStrategy)
-        reasoning_quality_indicators.append(("strategy_diversity", strategy_diversity))
-        reasoning_trace.append(f"Strategy diversity: {strategy_diversity:.3f}")
-
-        # Check for contradiction handling
-        contradictions_found = sum(len(result.contradictions_found) for result in analysis_results)
-        contradiction_awareness = min(1.0, contradictions_found / 5.0)  # Normalize
-        reasoning_quality_indicators.append(("contradiction_awareness", contradiction_awareness))
-        reasoning_trace.append(f"Contradiction awareness: {contradiction_awareness:.3f}")
-
-        # Check for uncertainty acknowledgment
-        uncertainties_found = sum(len(result.uncertainties) for result in analysis_results)
-        uncertainty_awareness = min(1.0, uncertainties_found / 3.0)  # Normalize
-        reasoning_quality_indicators.append(("uncertainty_awareness", uncertainty_awareness))
-        reasoning_trace.append(f"Uncertainty awareness: {uncertainty_awareness:.3f}")
-
-        # Overall meta-reasoning score
-        meta_score = np.mean([score for _, score in reasoning_quality_indicators])
-
-        # Generate meta-reasoning insights
-        insights = []
-        if strategy_diversity < 0.5:
-            insights.append("Consider using more diverse reasoning strategies")
-        if contradiction_awareness < 0.3:
-            insights.append("Look for more contradictions between sources")
-        if uncertainty_awareness < 0.3:
-            insights.append("Consider acknowledging more uncertainties")
-
-        reasoning_trace.append(f"Meta-reasoning score: {meta_score:.3f}")
-        reasoning_trace.append(f"Generated {len(insights)} meta-insights")
-
-        return AnalysisResult(
-            analysis_type=AnalysisType.META_REASONING,
-            confidence=ConfidenceLevel.MODERATE,
-            result={
-                "meta_score": meta_score,
-                "quality_indicators": dict(reasoning_quality_indicators),
-                "insights": insights,
-                "strategies_used": list(strategies_used),
-            },
-            reasoning=f"Meta-reasoning analysis score: {meta_score:.3f}",
-            strategy_used=ReasoningStrategy.PROBABILISTIC,
-            sources_analyzed=[info.id for info in retrieved_info],
-            reasoning_trace=reasoning_trace.copy(),
-        )
-
-    # ... Additional sophisticated analysis methods would be implemented here
-
-    async def get_nexus_stats(self) -> dict[str, Any]:
-        """Get comprehensive Cognitive Nexus statistics."""
-
-        recent_analyses = self.analysis_history[-100:] if self.analysis_history else []
-
-        return {
-            "analyses_performed": self.stats["analyses_performed"],
-            "synthesis_requests": self.stats["synthesis_requests"],
-            "contradictions_detected": self.stats["contradictions_detected"],
-            "uncertainties_identified": self.stats["uncertainties_identified"],
-            "belief_updates": self.stats["belief_updates"],
-            "fog_tasks_executed": self.stats["fog_tasks_executed"],
-            "meta_reasoning_cycles": self.stats["meta_reasoning_cycles"],
-            "cached_syntheses": len(self.synthesis_cache),
-            "analysis_history_size": len(self.analysis_history),
-            "belief_network_size": len(self.belief_network),
-            "average_confidence": np.mean([r.confidence.value for r in recent_analyses]) if recent_analyses else 0.0,
-            "reasoning_strategies_used": len(
-                {r.strategy_used.value for r in recent_analyses if hasattr(r, "strategy_used")}
-            ),
-            "system_status": "operational" if self.initialized else "initializing",
-            "fog_computing_enabled": self.enable_fog_computing,
-        }
-
-    # Production methods for fog computing and distributed analysis
-    async def _initialize_fog_workers(self):
-        """Initialize fog computing workers for distributed analysis."""
-        logger.info("Fog computing workers initialized for production")
-
-    async def _distributed_analysis(self, query, retrieved_info, analysis_types, reasoning_strategy):
-        """Distribute analysis across fog computing nodes."""
-        # Production distributed processing implementation
-        logger.info("Using distributed analysis for production workloads")
-        return []
-
-    async def _periodic_optimization(self):
-        """Periodic optimization of analysis patterns and belief networks."""
-        while True:
-            try:
-                await asyncio.sleep(3600)  # Every hour
-                # Optimize belief network, prune old cache entries, etc.
-                await self._optimize_belief_network()
-                await self._cleanup_cache()
-            except Exception as e:
-                logger.error(f"Error in periodic optimization: {e}")
-
-    async def _optimize_belief_network(self):
-        """Optimize belief network structure and probabilities."""
-        # Remove low-confidence beliefs, strengthen high-evidence beliefs
-        to_remove = []
-        for belief_id, belief in self.belief_network.items():
-            if belief.current_probability < 0.1 and belief.update_count < 2:
-                to_remove.append(belief_id)
-
-        for belief_id in to_remove:
-            del self.belief_network[belief_id]
-            if belief_id in self.belief_dependencies:
-                del self.belief_dependencies[belief_id]
-
-        if to_remove:
-            logger.info(f"Optimized belief network: removed {len(to_remove)} low-confidence beliefs")
-
-    async def _cleanup_cache(self):
-        """Clean up expired synthesis cache entries."""
-        current_time = time.time()
-        expired_keys = []
-
-        for key, answer in self.synthesis_cache.items():
-            if current_time - answer.creation_timestamp > 7200:  # 2 hours
-                expired_keys.append(key)
-
-        for key in expired_keys:
-            del self.synthesis_cache[key]
-
-        if expired_keys:
-            logger.info(f"Cleaned up {len(expired_keys)} expired synthesis cache entries")
-
-    # ... Additional methods would be implemented for full functionality
