--- core/rag/vector/contextual_vector_engine.py
@@ -1,920 +1,0 @@
-"""
-ContextualVectorEngine - High-Performance Vector Search with Dual Context Tags
-
-Advanced vector similarity search engine with contextual embeddings,
-dual context tags (book/chapter summaries), and semantic chunking.
-Optimized for high-performance retrieval with contextual awareness.
-
-This module provides the vector component of the unified HyperRAG system.
-"""
-
-import asyncio
-from dataclasses import dataclass, field
-from datetime import datetime
-from enum import Enum
-import logging
-import time
-from typing import Any
-import uuid
-
-import numpy as np
-
-logger = logging.getLogger(__name__)
-
-
-class ChunkingStrategy(Enum):
-    """Strategies for document chunking."""
-
-    FIXED_SIZE = "fixed_size"
-    SEMANTIC = "semantic"
-    SENTENCE = "sentence"
-    PARAGRAPH = "paragraph"
-    HIERARCHICAL = "hierarchical"
-
-
-class SimilarityMetric(Enum):
-    """Similarity metrics for vector comparison."""
-
-    COSINE = "cosine"
-    EUCLIDEAN = "euclidean"
-    DOT_PRODUCT = "dot_product"
-    MANHATTAN = "manhattan"
-
-
-@dataclass
-class ContextTag:
-    """Context tag providing additional information for retrieval."""
-
-    tag_type: str  # "book", "chapter", "section", "topic", etc.
-    content: str  # The actual context content
-    level: int  # Hierarchy level (0=highest, e.g., book level)
-    confidence: float = 1.0
-
-    # Metadata
-    created_at: datetime = field(default_factory=datetime.now)
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-@dataclass
-class VectorDocument:
-    """Document with vector embeddings and dual context tags."""
-
-    id: str = field(default_factory=lambda: str(uuid.uuid4()))
-    content: str = ""
-    title: str = ""
-
-    # Vector representations
-    content_embedding: np.ndarray | None = None
-    title_embedding: np.ndarray | None = None
-
-    # Dual context tags
-    primary_context: ContextTag | None = None  # e.g., book summary
-    secondary_context: ContextTag | None = None  # e.g., chapter summary
-    additional_contexts: list[ContextTag] = field(default_factory=list)
-
-    # Document metadata
-    doc_type: str = "text"
-    language: str = "en"
-    created_at: datetime = field(default_factory=datetime.now)
-
-    # Chunking information
-    chunk_index: int = 0
-    total_chunks: int = 1
-    parent_document_id: str | None = None
-
-    # Quality metrics
-    confidence_score: float = 1.0
-    relevance_score: float = 0.0  # Set during retrieval
-
-    # Metadata
-    tags: list[str] = field(default_factory=list)
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-    def get_combined_context(self) -> str:
-        """Get combined context from all context tags."""
-        contexts = []
-
-        if self.primary_context:
-            contexts.append(f"[{self.primary_context.tag_type}] {self.primary_context.content}")
-
-        if self.secondary_context:
-            contexts.append(f"[{self.secondary_context.tag_type}] {self.secondary_context.content}")
-
-        for context in self.additional_contexts:
-            contexts.append(f"[{context.tag_type}] {context.content}")
-
-        return " | ".join(contexts)
-
-    def calculate_context_similarity(self, query_context: str) -> float:
-        """Calculate similarity between document context and query context."""
-        doc_context = self.get_combined_context().lower()
-        query_context_lower = query_context.lower()
-
-        if not doc_context or not query_context_lower:
-            return 0.0
-
-        # Simple word overlap similarity (would use semantic similarity in production)
-        doc_words = set(doc_context.split())
-        query_words = set(query_context_lower.split())
-
-        if not doc_words or not query_words:
-            return 0.0
-
-        overlap = len(doc_words.intersection(query_words))
-        union = len(doc_words.union(query_words))
-
-        return overlap / union if union > 0 else 0.0
-
-
-@dataclass
-class VectorSearchResult:
-    """Result from vector similarity search."""
-
-    documents: list[VectorDocument] = field(default_factory=list)
-    similarities: list[float] = field(default_factory=list)
-
-    # Search metadata
-    query_time_ms: float = 0.0
-    total_candidates: int = 0
-    reranked: bool = False
-
-    # Context-aware results
-    context_matches: list[float] = field(default_factory=list)
-    combined_scores: list[float] = field(default_factory=list)
-
-    # Quality metrics
-    avg_similarity: float = 0.0
-    min_similarity: float = 0.0
-    max_similarity: float = 0.0
-
-    metadata: dict[str, Any] = field(default_factory=dict)
-
-
-class ContextualVectorEngine:
-    """
-    High-Performance Vector Search with Contextual Awareness
-
-    Advanced vector similarity search engine with dual context tags,
-    semantic chunking, and contextual embeddings. Designed for
-    high-performance retrieval with awareness of document structure
-    and hierarchical context.
-
-    Features:
-    - Dual context tags (book/chapter level summaries)
-    - Semantic chunking with overlap handling
-    - Multiple similarity metrics
-    - Context-aware reranking
-    - Hierarchical document organization
-    - Efficient vector indexing (FAISS with fallback to numpy)
-    """
-
-    def __init__(
-        self,
-        embedding_dim: int = 768,
-        similarity_threshold: float = 0.7,
-        enable_dual_context: bool = True,
-        enable_semantic_chunking: bool = True,
-        chunk_size: int = 512,
-        chunk_overlap: int = 50,
-    ):
-        self.embedding_dim = embedding_dim
-        self.similarity_threshold = similarity_threshold
-        self.enable_dual_context = enable_dual_context
-        self.enable_semantic_chunking = enable_semantic_chunking
-        self.chunk_size = chunk_size
-        self.chunk_overlap = chunk_overlap
-
-        # Document storage
-        self.documents: dict[str, VectorDocument] = {}
-
-        # Vector index (FAISS with numpy fallback)
-        self.vector_index = None
-        self.faiss_available = False
-        self.vector_matrix: np.ndarray | None = None
-        self.doc_id_mapping: list[str] = []  # Maps index position to doc_id
-
-        # Context indexing
-        self.context_index: dict[str, set[str]] = {}  # context_type -> doc_ids
-        self.tag_index: dict[str, set[str]] = {}  # tag -> doc_ids
-
-        # Caching
-        self.embedding_cache: dict[str, np.ndarray] = {}
-        self.search_cache: dict[str, VectorSearchResult] = {}
-
-        # Configuration
-        self.similarity_metric = SimilarityMetric.COSINE
-        self.chunking_strategy = ChunkingStrategy.SEMANTIC
-        self.rerank_with_context = True
-        self.cache_ttl = 3600  # 1 hour
-
-        # Statistics
-        self.stats = {
-            "documents_indexed": 0,
-            "chunks_created": 0,
-            "searches_performed": 0,
-            "cache_hits": 0,
-            "rerank_operations": 0,
-            "context_matches": 0,
-        }
-
-        self.initialized = False
-
-    async def initialize(self):
-        """Initialize the vector engine with optional FAISS support."""
-        logger.info("Initializing ContextualVectorEngine...")
-
-        # Try to initialize FAISS
-        try:
-            self.faiss_available = True
-            logger.info("✅ FAISS backend available for vector indexing")
-        except ImportError:
-            logger.warning("FAISS not available, using numpy fallback for vector operations")
-
-        # Initialize empty vector matrix
-        self.vector_matrix = np.empty((0, self.embedding_dim), dtype=np.float32)
-
-        self.initialized = True
-        logger.info("🔍 ContextualVectorEngine ready for high-performance retrieval")
-
-    async def index_document(
-        self,
-        content: str,
-        doc_id: str,
-        title: str = "",
-        primary_context: ContextTag | None = None,
-        secondary_context: ContextTag | None = None,
-        metadata: dict[str, Any] | None = None,
-    ) -> bool:
-        """Index a document with contextual embeddings."""
-        try:
-            # Create chunks if semantic chunking is enabled
-            if self.enable_semantic_chunking and len(content) > self.chunk_size:
-                chunks = await self._create_semantic_chunks(content, doc_id)
-            else:
-                chunks = [content]
-
-            documents_created = []
-
-            for i, chunk_content in enumerate(chunks):
-                # Create vector document
-                doc = VectorDocument(
-                    id=f"{doc_id}_chunk_{i}" if len(chunks) > 1 else doc_id,
-                    content=chunk_content,
-                    title=title,
-                    primary_context=primary_context,
-                    secondary_context=secondary_context,
-                    chunk_index=i,
-                    total_chunks=len(chunks),
-                    parent_document_id=doc_id if len(chunks) > 1 else None,
-                    metadata=metadata or {},
-                )
-
-                # Generate embeddings
-                doc.content_embedding = await self._create_embedding(chunk_content)
-                if title:
-                    doc.title_embedding = await self._create_embedding(title)
-
-                # Store document
-                self.documents[doc.id] = doc
-                documents_created.append(doc)
-
-                # Update context indexes
-                await self._update_context_indexes(doc)
-
-                self.stats["chunks_created"] += 1
-
-            # Add to vector index
-            await self._add_to_vector_index(documents_created)
-
-            self.stats["documents_indexed"] += 1
-            logger.debug(f"Indexed document {doc_id} with {len(chunks)} chunks")
-
-            return True
-
-        except Exception as e:
-            logger.exception(f"Failed to index document {doc_id}: {e}")
-            return False
-
-    async def search(
-        self,
-        query: str,
-        k: int = 10,
-        user_id: str | None = None,
-        context: dict[str, Any] | None = None,
-        similarity_threshold: float | None = None,
-    ) -> list[VectorDocument]:
-        """Perform contextual vector search."""
-        start_time = time.time()
-
-        try:
-            # Use provided threshold or default
-            threshold = similarity_threshold or self.similarity_threshold
-
-            # Check cache
-            cache_key = f"search:{hash(query)}:{k}:{threshold}:{user_id}"
-            if cache_key in self.search_cache:
-                self.stats["cache_hits"] += 1
-                return self.search_cache[cache_key].documents
-
-            # Generate query embedding
-            query_embedding = await self._create_embedding(query)
-
-            # Perform vector similarity search
-            candidates = await self._vector_search(query_embedding, k * 2, threshold)  # Get more candidates
-
-            # Context-aware reranking if enabled
-            if self.rerank_with_context and context:
-                candidates = await self._rerank_with_context(candidates, query, context)
-                self.stats["rerank_operations"] += 1
-
-            # Limit to requested number
-            final_results = candidates[:k]
-
-            # Update relevance scores
-            for i, doc in enumerate(final_results):
-                doc.relevance_score = 1.0 - (i / len(final_results))  # Decreasing relevance
-
-            # Create and cache result
-            query_time = (time.time() - start_time) * 1000
-            similarities = [doc.confidence_score for doc in final_results]
-
-            result = VectorSearchResult(
-                documents=final_results,
-                similarities=similarities,
-                query_time_ms=query_time,
-                total_candidates=len(candidates),
-                reranked=self.rerank_with_context and context is not None,
-                avg_similarity=np.mean(similarities) if similarities else 0.0,
-                min_similarity=min(similarities) if similarities else 0.0,
-                max_similarity=max(similarities) if similarities else 0.0,
-            )
-
-            self.search_cache[cache_key] = result
-            self.stats["searches_performed"] += 1
-
-            return final_results
-
-        except Exception as e:
-            logger.exception(f"Search failed: {e}")
-            return []
-
-    async def search_by_context(self, context_type: str, context_content: str, k: int = 10) -> list[VectorDocument]:
-        """Search documents by context tags."""
-        try:
-            # Get documents with matching context type
-            candidate_ids = self.context_index.get(context_type, set())
-
-            if not candidate_ids:
-                return []
-
-            # Calculate context similarity for each candidate
-            candidates_with_scores = []
-            for doc_id in candidate_ids:
-                doc = self.documents[doc_id]
-                context_sim = doc.calculate_context_similarity(context_content)
-
-                if context_sim > 0.1:  # Minimum context similarity
-                    candidates_with_scores.append((doc, context_sim))
-
-            # Sort by context similarity
-            candidates_with_scores.sort(key=lambda x: x[1], reverse=True)
-
-            # Return top k documents
-            results = [doc for doc, _ in candidates_with_scores[:k]]
-
-            self.stats["context_matches"] += len(results)
-            return results
-
-        except Exception as e:
-            logger.exception(f"Context search failed: {e}")
-            return []
-
-    async def get_similar_documents(
-        self, doc_id: str, k: int = 5, exclude_same_parent: bool = True
-    ) -> list[tuple[VectorDocument, float]]:
-        """Find documents similar to a given document."""
-        try:
-            if doc_id not in self.documents:
-                return []
-
-            source_doc = self.documents[doc_id]
-            if source_doc.content_embedding is None:
-                return []
-
-            # Perform similarity search using document's embedding
-            candidates = await self._vector_search(
-                source_doc.content_embedding,
-                k + 10,  # Get extra to account for filtering
-                0.0,  # No threshold for similarity search
-            )
-
-            # Filter out the source document and same parent if requested
-            filtered_candidates = []
-            for doc in candidates:
-                if doc.id == doc_id:
-                    continue
-
-                if (
-                    exclude_same_parent
-                    and source_doc.parent_document_id
-                    and doc.parent_document_id == source_doc.parent_document_id
-                ):
-                    continue
-
-                # Calculate similarity score
-                similarity = await self._calculate_similarity(source_doc.content_embedding, doc.content_embedding)
-
-                filtered_candidates.append((doc, similarity))
-
-            # Sort by similarity and return top k
-            filtered_candidates.sort(key=lambda x: x[1], reverse=True)
-            return filtered_candidates[:k]
-
-        except Exception as e:
-            logger.exception(f"Similar document search failed: {e}")
-            return []
-
-    async def get_status(self) -> dict[str, Any]:
-        """Get status and performance metrics."""
-        try:
-            # Calculate index statistics
-            total_docs = len(self.documents)
-            total_contexts = len(self.context_index)
-            total_tags = len(self.tag_index)
-
-            # Vector index statistics
-            vector_index_size = self.vector_matrix.shape[0] if self.vector_matrix is not None else 0
-
-            # Cache statistics
-            cache_hit_rate = self.stats["cache_hits"] / max(1, self.stats["searches_performed"])
-
-            # Memory usage estimation
-            embedding_memory_mb = total_docs * self.embedding_dim * 4 / (1024 * 1024)  # 4 bytes per float32
-
-            return {
-                "status": "healthy",
-                "index_size": {
-                    "total_documents": total_docs,
-                    "vector_index_size": vector_index_size,
-                    "context_types": total_contexts,
-                    "unique_tags": total_tags,
-                },
-                "performance": {
-                    "cache_hit_rate": cache_hit_rate,
-                    "faiss_available": self.faiss_available,
-                    "similarity_metric": self.similarity_metric.value,
-                    "chunking_strategy": self.chunking_strategy.value,
-                },
-                "memory": {
-                    "estimated_embedding_memory_mb": embedding_memory_mb,
-                    "cache_entries": len(self.search_cache),
-                },
-                "configuration": {
-                    "embedding_dim": self.embedding_dim,
-                    "similarity_threshold": self.similarity_threshold,
-                    "dual_context_enabled": self.enable_dual_context,
-                    "semantic_chunking_enabled": self.enable_semantic_chunking,
-                    "chunk_size": self.chunk_size,
-                    "chunk_overlap": self.chunk_overlap,
-                },
-                "statistics": self.stats.copy(),
-            }
-
-        except Exception as e:
-            logger.exception(f"Status check failed: {e}")
-            return {"status": "error", "error": str(e)}
-
-    async def close(self):
-        """Close the vector engine and clean up resources."""
-        logger.info("Closing ContextualVectorEngine...")
-
-        # Clear all data structures
-        self.documents.clear()
-        self.context_index.clear()
-        self.tag_index.clear()
-        self.embedding_cache.clear()
-        self.search_cache.clear()
-
-        # Clear vector index
-        self.vector_matrix = None
-        self.doc_id_mapping.clear()
-
-        logger.info("ContextualVectorEngine closed")
-
-    # Private implementation methods
-
-    async def _create_embedding(self, text: str) -> np.ndarray:
-        """Create embedding for text (with caching)."""
-        # Check cache first
-        text_hash = str(hash(text))
-        if text_hash in self.embedding_cache:
-            return self.embedding_cache[text_hash]
-
-        try:
-            # Create deterministic pseudo-embedding from text hash
-            # In production, this would use a real embedding model
-            import hashlib
-
-            text_bytes = text.encode("utf-8")
-            text_hash_bytes = hashlib.md5(text_bytes, usedforsecurity=False).hexdigest()
-            seed = int(text_hash_bytes[:8], 16)
-
-            np.random.seed(seed)
-
-            # Generate normalized embedding
-            embedding = np.random.normal(0, 1, self.embedding_dim).astype(np.float32)
-            embedding = embedding / (np.linalg.norm(embedding) + 1e-8)
-
-            # Cache the result
-            self.embedding_cache[text_hash] = embedding
-
-            return embedding
-
-        except Exception as e:
-            logger.warning(f"Embedding generation failed: {e}")
-            # Return zero vector as fallback
-            return np.zeros(self.embedding_dim, dtype=np.float32)
-
-    async def _create_semantic_chunks(self, content: str, doc_id: str) -> list[str]:
-        """Create semantic chunks from content."""
-        # Simple sentence-based chunking (would use more sophisticated methods in production)
-        sentences = content.split(". ")
-        chunks = []
-        current_chunk = ""
-
-        for sentence in sentences:
-            # Add sentence to current chunk
-            test_chunk = current_chunk + ". " + sentence if current_chunk else sentence
-
-            # Check if chunk size would exceed limit
-            if len(test_chunk) > self.chunk_size and current_chunk:
-                # Add overlap from current chunk
-                overlap_words = current_chunk.split()[-self.chunk_overlap :]
-                chunks.append(current_chunk)
-                current_chunk = " ".join(overlap_words) + ". " + sentence
-            else:
-                current_chunk = test_chunk
-
-        # Add the last chunk
-        if current_chunk:
-            chunks.append(current_chunk)
-
-        # Ensure we have at least one chunk
-        if not chunks:
-            chunks = [content]
-
-        return chunks
-
-    async def _update_context_indexes(self, doc: VectorDocument):
-        """Update context and tag indexes for a document."""
-        # Index primary context
-        if doc.primary_context:
-            context_type = doc.primary_context.tag_type
-            if context_type not in self.context_index:
-                self.context_index[context_type] = set()
-            self.context_index[context_type].add(doc.id)
-
-        # Index secondary context
-        if doc.secondary_context:
-            context_type = doc.secondary_context.tag_type
-            if context_type not in self.context_index:
-                self.context_index[context_type] = set()
-            self.context_index[context_type].add(doc.id)
-
-        # Index additional contexts
-        for context in doc.additional_contexts:
-            context_type = context.tag_type
-            if context_type not in self.context_index:
-                self.context_index[context_type] = set()
-            self.context_index[context_type].add(doc.id)
-
-        # Index tags
-        for tag in doc.tags:
-            if tag not in self.tag_index:
-                self.tag_index[tag] = set()
-            self.tag_index[tag].add(doc.id)
-
-    async def _add_to_vector_index(self, documents: list[VectorDocument]):
-        """Add documents to the vector index."""
-        # Extract embeddings
-        embeddings = []
-        doc_ids = []
-
-        for doc in documents:
-            if doc.content_embedding is not None:
-                embeddings.append(doc.content_embedding)
-                doc_ids.append(doc.id)
-
-        if not embeddings:
-            return
-
-        # Convert to numpy array
-        new_embeddings = np.array(embeddings, dtype=np.float32)
-
-        # Add to vector matrix
-        if self.vector_matrix.shape[0] == 0:
-            self.vector_matrix = new_embeddings
-        else:
-            self.vector_matrix = np.vstack([self.vector_matrix, new_embeddings])
-
-        # Update document ID mapping
-        self.doc_id_mapping.extend(doc_ids)
-
-        # Build/update FAISS index if available
-        try:
-            self._build_faiss_index()
-        except ImportError:
-            logger.info("FAISS not available, using numpy-based search")
-
-    async def _vector_search(self, query_embedding: np.ndarray, k: int, threshold: float) -> list[VectorDocument]:
-        """Perform vector similarity search."""
-        if self.vector_matrix.shape[0] == 0:
-            return []
-
-        try:
-            # Calculate similarities
-            similarities = await self._calculate_similarities(query_embedding, self.vector_matrix)
-
-            # Get top k indices above threshold
-            valid_indices = np.where(similarities >= threshold)[0]
-            if len(valid_indices) == 0:
-                return []
-
-            # Sort by similarity (descending)
-            sorted_indices = valid_indices[np.argsort(similarities[valid_indices])[::-1]]
-            top_indices = sorted_indices[:k]
-
-            # Get corresponding documents
-            results = []
-            for idx in top_indices:
-                doc_id = self.doc_id_mapping[idx]
-                doc = self.documents[doc_id]
-                doc.confidence_score = similarities[idx]  # Store similarity as confidence
-                results.append(doc)
-
-            return results
-
-        except Exception as e:
-            logger.exception(f"Vector search failed: {e}")
-            return []
-
-    async def _calculate_similarities(self, query_embedding: np.ndarray, index_embeddings: np.ndarray) -> np.ndarray:
-        """Calculate similarity scores between query and index embeddings."""
-        if self.similarity_metric == SimilarityMetric.COSINE:
-            # Cosine similarity
-            similarities = np.dot(index_embeddings, query_embedding)
-            return similarities
-
-        elif self.similarity_metric == SimilarityMetric.EUCLIDEAN:
-            # Euclidean distance (converted to similarity)
-            distances = np.linalg.norm(index_embeddings - query_embedding, axis=1)
-            max_distance = np.max(distances)
-            similarities = 1.0 - (distances / max_distance)
-            return similarities
-
-        elif self.similarity_metric == SimilarityMetric.DOT_PRODUCT:
-            # Dot product
-            similarities = np.dot(index_embeddings, query_embedding)
-            return similarities
-
-        else:
-            # Default to cosine similarity
-            similarities = np.dot(index_embeddings, query_embedding)
-            return similarities
-
-    async def _calculate_similarity(self, embedding1: np.ndarray, embedding2: np.ndarray) -> float:
-        """Calculate similarity between two embeddings."""
-        if self.similarity_metric == SimilarityMetric.COSINE:
-            return float(np.dot(embedding1, embedding2))
-        elif self.similarity_metric == SimilarityMetric.EUCLIDEAN:
-            distance = np.linalg.norm(embedding1 - embedding2)
-            return float(1.0 / (1.0 + distance))  # Convert distance to similarity
-        elif self.similarity_metric == SimilarityMetric.DOT_PRODUCT:
-            return float(np.dot(embedding1, embedding2))
-        else:
-            return float(np.dot(embedding1, embedding2))
-
-    async def _rerank_with_context(
-        self, candidates: list[VectorDocument], query: str, context: dict[str, Any]
-    ) -> list[VectorDocument]:
-        """Rerank candidates using contextual information."""
-        try:
-            query_context = context.get("context", "")
-            user_preferences = context.get("user_preferences", {})
-
-            # Calculate context-aware scores
-            reranked_candidates = []
-            for doc in candidates:
-                # Base similarity score
-                base_score = doc.confidence_score
-
-                # Context similarity boost
-                context_sim = doc.calculate_context_similarity(query_context)
-                context_boost = context_sim * 0.3  # 30% weight for context
-
-                # User preference boost (simplified)
-                preference_boost = 0.0
-                for pref_key, pref_value in user_preferences.items():
-                    if pref_key in doc.metadata and doc.metadata[pref_key] == pref_value:
-                        preference_boost += 0.1
-
-                # Combined score
-                final_score = base_score + context_boost + preference_boost
-                doc.confidence_score = min(1.0, final_score)
-
-                reranked_candidates.append(doc)
-
-            # Sort by final score
-            reranked_candidates.sort(key=lambda x: x.confidence_score, reverse=True)
-            return reranked_candidates
-
-        except Exception as e:
-            logger.warning(f"Context reranking failed: {e}")
-            return candidates
-
-    def _build_faiss_index(self):
-        """Build FAISS index for fast vector similarity search."""
-        try:
-            import faiss
-
-            if self.vector_matrix is None or self.vector_matrix.shape[0] == 0:
-                logger.info("No vectors to index")
-                return
-
-            # Normalize vectors for cosine similarity if needed
-            if self.similarity_metric == SimilarityMetric.COSINE:
-                # Normalize vectors to unit length for cosine similarity
-                faiss.normalize_L2(self.vector_matrix)
-
-            # Create appropriate FAISS index based on vector size
-            vector_dim = self.vector_matrix.shape[1]
-
-            if self.vector_matrix.shape[0] < 1000:
-                # For small datasets, use flat index
-                self.vector_index = faiss.IndexFlatIP(vector_dim)  # Inner product (cosine after normalization)
-            else:
-                # For larger datasets, use IVF index with clustering
-                nlist = min(100, self.vector_matrix.shape[0] // 10)  # Number of clusters
-                quantizer = faiss.IndexFlatIP(vector_dim)
-                self.vector_index = faiss.IndexIVFFlat(quantizer, vector_dim, nlist)
-
-                # Train the index (required for IVF indices)
-                self.vector_index.train(self.vector_matrix.astype(np.float32))
-
-            # Add vectors to index
-            self.vector_index.add(self.vector_matrix.astype(np.float32))
-            self.faiss_available = True
-
-            logger.info(f"Built FAISS index with {self.vector_index.ntotal} vectors")
-
-        except ImportError:
-            logger.info("FAISS not available, using numpy-based search")
-            self.faiss_available = False
-        except Exception as e:
-            logger.error(f"Failed to build FAISS index: {e}")
-            self.faiss_available = False
-
-    async def _faiss_search(self, query_embedding: np.ndarray, k: int) -> tuple[list[float], list[int]]:
-        """Perform fast similarity search using FAISS."""
-        if not self.faiss_available or self.vector_index is None:
-            return [], []
-
-        try:
-            import faiss
-
-            # Normalize query if using cosine similarity
-            if self.similarity_metric == SimilarityMetric.COSINE:
-                query_norm = query_embedding.copy().reshape(1, -1).astype(np.float32)
-                faiss.normalize_L2(query_norm)
-            else:
-                query_norm = query_embedding.reshape(1, -1).astype(np.float32)
-
-            # Search
-            scores, indices = self.vector_index.search(query_norm, k)
-
-            # Convert to lists and filter valid results
-            scores_list = scores[0].tolist()
-            indices_list = indices[0].tolist()
-
-            # Filter out invalid indices (-1 indicates no match)
-            valid_results = [(score, idx) for score, idx in zip(scores_list, indices_list) if idx >= 0]
-
-            if valid_results:
-                scores, indices = zip(*valid_results)
-                return list(scores), list(indices)
-            else:
-                return [], []
-
-        except Exception as e:
-            logger.error(f"FAISS search failed: {e}")
-            return [], []
-
-
-# Factory functions for creating contextual components
-
-
-def create_context_tag(tag_type: str, content: str, level: int = 0, confidence: float = 1.0) -> ContextTag:
-    """Create a context tag for dual context system."""
-    return ContextTag(
-        tag_type=tag_type,
-        content=content,
-        level=level,
-        confidence=confidence,
-        metadata={"created_by": "contextual_vector_engine"},
-    )
-
-
-def create_book_chapter_contexts(
-    book_title: str, book_summary: str, chapter_title: str, chapter_summary: str
-) -> tuple[ContextTag, ContextTag]:
-    """Create book and chapter context tags."""
-    book_context = create_context_tag(tag_type="book", content=f"{book_title}: {book_summary}", level=0, confidence=1.0)
-
-    chapter_context = create_context_tag(
-        tag_type="chapter", content=f"{chapter_title}: {chapter_summary}", level=1, confidence=1.0
-    )
-
-    return book_context, chapter_context
-
-
-if __name__ == "__main__":
-
-    async def test_contextual_vector_engine():
-        """Test ContextualVectorEngine functionality."""
-        # Create system
-        engine = ContextualVectorEngine(
-            embedding_dim=384, similarity_threshold=0.7, enable_dual_context=True, enable_semantic_chunking=True
-        )
-        await engine.initialize()
-
-        # Create context tags
-        book_context, chapter_context = create_book_chapter_contexts(
-            book_title="Machine Learning Fundamentals",
-            book_summary="Comprehensive guide to machine learning concepts and algorithms",
-            chapter_title="Neural Networks",
-            chapter_summary="Introduction to artificial neural networks and deep learning",
-        )
-
-        # Index a document
-        success = await engine.index_document(
-            content="Neural networks are computational models inspired by biological neural networks. They consist of interconnected nodes called neurons that process information.",
-            doc_id="ml_book_chapter3_section1",
-            title="Introduction to Neural Networks",
-            primary_context=book_context,
-            secondary_context=chapter_context,
-            metadata={"author": "AI Researcher", "year": 2024},
-        )
-        print(f"Document indexed: {success}")
-
-        # Test vector search
-        results = await engine.search(
-            query="neural networks artificial intelligence", k=5, context={"context": "machine learning deep learning"}
-        )
-        print(f"Vector search found {len(results)} results")
-
-        # Test context search
-        context_results = await engine.search_by_context(context_type="book", context_content="machine learning", k=3)
-        print(f"Context search found {len(context_results)} results")
-
-        # Test similar documents
-        if results:
-            similar_docs = await engine.get_similar_documents(doc_id=results[0].id, k=3)
-            print(f"Found {len(similar_docs)} similar documents")
-
-        # Status check
-        status = await engine.get_status()
-        print(f"Engine status: {status['status']}")
-        print(f"Index size: {status['index_size']}")
-        print(f"Performance: {status['performance']}")
-
-        await engine.close()
-
-    import asyncio
-
-    asyncio.run(test_contextual_vector_engine())
-
-
-@dataclass
-class ContextualChunk:
-    """A chunk of content with contextual information."""
-
-    chunk_id: str = field(default_factory=lambda: str(uuid.uuid4()))
-    content: str = ""
-    start_index: int = 0
-    end_index: int = 0
-
-    # Context information
-    book_context: str | None = None
-    chapter_context: str | None = None
-    section_context: str | None = None
-
-    # Vector representation
-    embedding: np.ndarray | None = None
-
-    # Quality metrics
-    semantic_coherence: float = 1.0
-    information_density: float = 1.0
-
-    # Metadata
-    metadata: dict[str, Any] = field(default_factory=dict)
-    created_at: datetime = field(default_factory=datetime.now)
